Adds message for succesfull sign in. Moves country list to Modules folder.
