+++ src/Components/LoginForm.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import { Form, Button } from 'semantic-ui-react'
+import { Form, Button, Container, Message } from 'semantic-ui-react'
 import { connect } from 'react-redux'
 import { signInUser } from '../reduxTokenAuthConfig'
 
@@ -7,7 +7,8 @@ class LoginForm extends Component {
   state = {
     email: '',
     password: '',
-    errors: ''
+    errors: '',
+    message: false
   }
 
   onChangeHandler = (e) => {
@@ -25,32 +26,50 @@ class LoginForm extends Component {
     } = this.state
     try {
       await signInUser({ email, password })
-      setTimeout(function(){ history.push('/') }, 3000)
+      this.setState({ message: true })
+      setTimeout(function () { history.push('/') }, 3000)
     } catch (response) { console.log(response.errors) }
   }
 
   render() {
 
+    let message
+
+    if (this.state.message === true) {
+      message = (
+        <>
+          <br />
+          <Message color="green">
+            <p>You have succesfully logged in. Wait to be redirected to the main page.</p>
+          </Message>
+        </>
+      )
+    }
+
     return (
-      <Form id="login-form" onSubmit={this.onSubmit}>
-        <Form.Input
-          id="email"
-          value={this.state.email}
-          onChange={this.onChangeHandler}
-          placeholder="Email"
-        />
+      <Container>
+        <p>{message}</p>
+        <Form id="login-form" onSubmit={this.onSubmit}>
+          <Form.Input
+            id="email"
+            value={this.state.email}
+            onChange={this.onChangeHandler}
+            placeholder="Email"
+          />
 
-        <Form.Input
-          id="password"
-          type="password"
-          value={this.state.password}
-          onChange={this.onChangeHandler}
-          placeholder="Password"
-        />
+          <Form.Input
+            id="password"
+            type="password"
+            value={this.state.password}
+            onChange={this.onChangeHandler}
+            placeholder="Password"
+          />
 
-        <Button id="login">Login</Button>
+          <Button id="login">Login</Button>
 
-      </Form>
+        </Form>
+
+      </Container>
     )
   }
